Move grammar out of compiler
